refactor<ProductUpdatingService> 상품검색 통합
기존 ProductDetailService, ProductSearchService 분리 되있던 부분 ProductSearchService로 통합
