+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/ProductService.java
@@ -1,18 +1,16 @@
 package com.gulbi.Backend.domain.rental.product.service.product;
 
 import com.gulbi.Backend.domain.rental.product.dto.product.ProductOverViewResponse;
-import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductSearchRequestDto;
+import com.gulbi.Backend.domain.rental.product.dto.product.request.*;
 import com.gulbi.Backend.domain.rental.product.dto.product.response.ProductDetailResponseDto;
-import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductRegisterRequestDto;
-import com.gulbi.Backend.domain.rental.product.entity.Product;
+import com.gulbi.Backend.domain.rental.product.vo.image.ProductImageCollection;
 
-import java.io.IOException;
 import java.util.List;
 
 public interface ProductService {
     public List<ProductOverViewResponse> searchProductOverview(ProductSearchRequestDto productSearchRequestDto);
-    public void registrationProduct(ProductRegisterRequestDto productRegisterRequestDto);
+    public void registrationProduct(ProductRegisterRequestDto productRegisterRequestDto, ProductImageCreateRequestDto productImageCreateRequestDto);
     public ProductDetailResponseDto getProductDetail(Long productId);
     public void updateProductViews(Long productId);
-
+    public void updateProduct(ProductUpdateRequestDto productUpdateRequestDto, ProductCategoryUpdateRequestDto productCategoryUpdateRequestDto, ProductImageDeleteRequestDto productImageDeleteRequestDto,ProductImageCreateRequestDto productImageCreateRequestDto);
 }
